fix: improve file handling by preventing memoization issues, providing config values at run time
